Reviung: Add experimental second config to test reviung underglow (buggy atm)
I'm using 42keebs frood (pro micro rp2040 compatible), and by default reviung
goes not support underflow for this.

At first I tried to add a board config to a local boards/shields/reviung41/boards/sparkfun_pro_micro_rp2040.overlay,
but the build did not pick that up (I did remember to add it to git).
As a brute force method, I added a "fork" of the whole shield, and
lazily renamed it by an added "z".

The underglow works and lights up, and the rgb buttons seems to do something,
but the light is flickering when typing. This is no good, but as I run
out of time for now, I'll just commit this hopefully temporary testing
shield config for later use.